Use modern dependencies (#9)
* Re-relax bundle dependencies

We will now make the gem re-compatible with the latest versions.

* Update controller specs for Rails 6

* Parameters in controller specs must now be wrapped in a hash
  supplied with the `params` key
* ActionController::TestResponse no longer responds to `success?` but
  rather `successful?`
* Rendering a plain string is done via `render plain:` now.

* Sprockets 4 expects a manifest.js file

It doesn't actually matter if this file is empty or not, but it must
be present.

* Sprockets::Index is gone in Sprockets 4

However, this object may now be a `Sprockets::CachedEnvironment`
instead.

* Include image methods in Sprockets 4 processor

It's no longer sufficient to include these methods in `Sass::Script`.

* Sprockets::Asset no longer provides `pathname` method

Instead, we are directed to use `filename`, but other uses in this
library expect to be able to call `read` on the return object, so we
must return a `Pathname` instance.

* Test against modern Ruby versions

* Use SassC instead of Sass gem

Ruby Sass is deprecated; see https://sass-lang.com/ruby-sass for
details. It was deprecated on 26th March 2019, and the recommended
migration path for most Ruby users is to SassC, which is built on top
of libsass.